+++ idearium-lib/test/scripts/test.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env sh
 
-DEBUG=lib:logger:* BLUEBIRD_WARNINGS=0 "/app/node_modules/.bin/_mocha"
+DEBUG=lib:logger:*,idearium-lib:common:exception BLUEBIRD_WARNINGS=0 "/app/node_modules/.bin/_mocha"