+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This file is a history of the changes made to idearium-lib.
 ## Unreleased
 
 - Simplified `common/exception` now that Opbeat should become standard practice.
+- Added a `query` file to help create performant queries.
+- Added `kue` and `queue` to the common files.
 
 ## 1.0.0-alpha.13