+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This file is a history of the changes made to idearium-lib.
 
-## Unreleased
+## 1.0.0-alpha.14
 
 - Simplified `common/exception` now that Opbeat should become standard practice.
 - Added a `query` file to help create performant queries.